shift testing files around
